# api/controller/role_controller.go
package controller

import (
	"errors"
	"net/http"
	"shadmin/domain"
	"shadmin/internal/casbin"
	"shadmin/internal/conf"

	"github.com/gin-gonic/gin"
)

type RoleController struct {
	CasManager     casbin.Manager
	Env            *conf.Env
	RoleUseCase    domain.RoleUseCase
	UserRepository domain.UserRepository
	RoleRepository domain.RoleRepository
	MenuRepository domain.MenuRepository
}

func (rc *RoleController) getRoleIDOrBadRequest(c *gin.Context) (string, bool) {
	roleID := c.Param("id")
	if roleID == "" {
		c.JSON(http.StatusBadRequest, domain.RespError("Role ID is required"))
		return "", false
	}

	return roleID, true
}

func (rc *RoleController) writeRoleMutationError(c *gin.Context, err error, forbiddenErr error) {
	if errors.Is(err, forbiddenErr) {
		c.JSON(http.StatusForbidden, domain.RespError(err.Error()))
		return
	}
	if err.Error() == "role not found" {
		c.JSON(http.StatusNotFound, domain.RespError(err.Error()))
		return
	}

	c.JSON(http.StatusInternalServerError, domain.RespError(err.Error()))
}

// RoleInfo 角色信息结构
type RoleInfo struct {
	ID   string `json:"id"`   // 角色ID
	Name string `json:"name"` // 角色显示名称
	Type string `json:"type"` // 角色类型
}

// GetRoles 获取所有角色
// @Summary      Get all roles
// @Description  Retrieve a list of all available roles in the system
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object} domain.Response{data=[]RoleInfo}  "Successfully retrieved roles"
// @Failure      500  {object} domain.Response  "Internal server error"
// @Router       /system/role [get]
func (rc *RoleController) GetRoles(c *gin.Context) {
	// 获取所有角色
	roles, err := rc.RoleUseCase.Fetch(c.Request.Context())
	if err != nil {
		c.JSON(http.StatusInternalServerError, domain.RespError(err.Error()))
		return
	}

	c.JSON(http.StatusOK, domain.RespSuccess(roles))
}

// CreateRole 创建自定义角色
// @Summary      Create a new custom role
// @Description  Create a new custom role in the system
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body     domain.CreateRoleRequest  true  "Role creation information"
// @Success      201      {object} domain.Response{data=domain.Role}  "Successfully created role"
// @Failure      400      {object} domain.Response  "Bad request - invalid parameters"
// @Failure      409      {object} domain.Response  "Role code already exists"
// @Failure      500      {object} domain.Response  "Internal server error"
// @Router       /system/role [post]
func (rc *RoleController) CreateRole(c *gin.Context) {
	var request domain.CreateRoleRequest
	if !MustBindJSON(c, &request) {
		return
	}

	// Set default status if not provided
	if request.Status == "" {
		request.Status = domain.RoleStatusActive
	}

	// Create the role
	err := rc.RoleUseCase.Create(c.Request.Context(), &request)
	if err != nil {
		if err.Error() == "role code already exists" {
			c.JSON(http.StatusConflict, domain.RespError(err.Error()))
			return
		}
		c.JSON(http.StatusInternalServerError, domain.RespError(err.Error()))
		return
	}

	c.JSON(http.StatusCreated, domain.RespSuccess(err))
}

// getRoleByID extracts the role ID from the path, fetches the role, and writes
// an error response when something goes wrong. Returns nil if a response was already sent.
func (rc *RoleController) getRoleByID(c *gin.Context) *domain.Role {
	roleID, ok := rc.getRoleIDOrBadRequest(c)
	if !ok {
		return nil
	}

	role, err := rc.RoleUseCase.GetByID(c.Request.Context(), roleID)
	if err != nil {
		if err.Error() == "role not found" {
			c.JSON(http.StatusNotFound, domain.RespError(err.Error()))
			return nil
		}
		c.JSON(http.StatusInternalServerError, domain.RespError(err.Error()))
		return nil
	}

	return role
}

// GetRole 获取角色详情
// @Summary      Get role by ID
// @Description  Get a specific role by its ID
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path     string  true  "Role ID"
// @Success      200  {object} domain.Response{data=domain.Role}  "Successfully retrieved role"
// @Failure      404  {object} domain.Response  "Role not found"
// @Failure      500  {object} domain.Response  "Internal server error"
// @Router       /system/role/{id} [get]
func (rc *RoleController) GetRole(c *gin.Context) {
	role := rc.getRoleByID(c)
	if role == nil {
		return
	}

	c.JSON(http.StatusOK, domain.RespSuccess(role))
}

// UpdateRole 更新角色
// @Summary      Update role
// @Description  Update an existing role
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id       path     string                     true  "Role ID"
// @Param        request  body     domain.UpdateRoleRequest   true  "Role update information"
// @Success      200      {object} domain.Response{data=domain.Role}  "Successfully updated role"
// @Failure      400      {object} domain.Response  "Bad request - invalid parameters"
// @Failure      404      {object} domain.Response  "Role not found"
// @Failure      409      {object} domain.Response  "Role code already exists"
// @Failure      500      {object} domain.Response  "Internal server error"
// @Router       /system/role/{id} [put]
func (rc *RoleController) UpdateRole(c *gin.Context) {
	roleID, ok := rc.getRoleIDOrBadRequest(c)
	if !ok {
		return
	}

	var request domain.UpdateRoleRequest
	if !MustBindJSON(c, &request) {
		return
	}

	role, err := rc.RoleUseCase.Update(c.Request.Context(), roleID, &request)
	if err != nil {
		rc.writeRoleMutationError(c, err, domain.ErrCannotRenameAdminRole)
		return
	}

	c.JSON(http.StatusOK, domain.RespSuccess(role))
}

// DeleteRole 删除角色
// @Summary      Delete role
// @Description  Delete an existing role
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path     string  true  "Role ID"
// @Success      200  {object} domain.Response  "Successfully deleted role"
// @Failure      404  {object} domain.Response  "Role not found"
// @Failure      500  {object} domain.Response  "Internal server error"
// @Router       /system/role/{id} [delete]
func (rc *RoleController) DeleteRole(c *gin.Context) {
	roleID, ok := rc.getRoleIDOrBadRequest(c)
	if !ok {
		return
	}

	err := rc.RoleUseCase.Delete(c.Request.Context(), roleID)
	if err != nil {
		rc.writeRoleMutationError(c, err, domain.ErrCannotDeleteAdminRole)
		return
	}

	// 注意：由于使用集中同步模式，角色关系会在下次同步时自动清理
	// 不需要手动删除 Casbin 中的角色分配
	c.JSON(http.StatusOK, domain.RespSuccess("Role deleted successfully"))
}

// GetRoleMenus 获取角色的菜单权限
// @Summary      Get role menu permissions
// @Description  Get menu permissions assigned to a role
// @Tags         Roles
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id   path     string  true  "Role ID"
// @Success      200  {object} domain.Response{data=[]string}  "Successfully retrieved role menu permissions"
// @Failure      400  {object} domain.Response  "Bad request - Role ID is required"
// @Failure      404  {object} domain.Response  "Role not found"
// @Failure      500  {object} domain.Response  "Internal server error"
// @Router       /system/role/menus/{id} [get]
func (rc *RoleController) GetRoleMenus(c *gin.Context) {
	role := rc.getRoleByID(c)
	if role == nil {
		return
	}

	c.JSON(http.StatusOK, domain.RespSuccess(role.MenusIds))
}